% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_LCP.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Is a character string giving the file name of the input data
file.}
}
\value{
If a file name is provided this functions returns a dplyr version of
  an R data frame with the data contained in filename. Otherwise it stops.
}
\description{
A read data file function.

\code{fars_read} This function reads a filename from the Fatality Analysis
Reporting System (FARS) into a dplyr version of an R data frame. It stops if
the file does not exist.
}
\examples{
fars_read("accident_2015.csv.bz2")

\dontrun{
fars_read()
}

}
\references{
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
