% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_LCP.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Is a numeric vector corresponding to the series of years for
which you want to extract monthly data.}
}
\value{
A list of subsetted dataframes in dplyr format with the monthly data from
  every year included in years.
}
\description{
A function to extraxct monthly data from a given set of years.

\code{fars_read_years} This function uses an user provided series of years to
extract their FARS monthly data
}
\details{
For each year provided in the numeric vector argument to the
  function, the make_filename function from this package generates its
  corresponding filename. The data from each file will then be extracted
  using the fars_read function from this package. Subsequently, the monthly
  data is subsetted by means of the dplyr functions mutate and select. The
  functions evaluates the code and assigns a warning handler to years for
  which data are not available.
}
\examples{
fars_read_years(c(2013,2014,2015))

\dontrun{
fars_read_years(c(2016,2017))
}

}
\references{
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
