% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_LCP.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Is a numeric vector corresponding to a series of years for which
to summarize FARS data.}
}
\value{
A dataframe with the summary of fars events per month for every given
  year.
}
\description{
A function to summarize FARS events per month and year.

\code{fars_summarize_years} This function uses an user provided series of
years to create a summary of the total number of FARS events per month and
year.
}
\examples{
fars_summarize_years(c(2013,2014,2015))

\dontrun{
fars_summarize_years(c(2016,2017))
}

}
\references{
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
