% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_LCP.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric corresponding to the user provided year.}
}
\value{
A character string corrresponding to the full file name with the
  format. accident_year.csv.bz2.
}
\description{
A simple function to create the filename of accident data for a given year.

\code{make_filename} This function takes an user provided year to generate a
filename from FARS.
}
\details{
The function explictly converts the numeric to an integer.
}
\examples{
make_filename(2013)

}
\references{
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
